separate tests in pkg/object
