+++ components/eventing-controller/pkg/object/equality_test.go
@@ -1,231 +1,18 @@
-package object
+//go:build unit
+// +build unit
+
+package object_test
 
 import (
-	"net/http"
 	"testing"
 
-	apigatewayv1beta1 "github.com/kyma-incubator/api-gateway/api/v1beta1"
-	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	eventingv1alpha1 "github.com/kyma-project/kyma/components/eventing-controller/api/v1alpha1"
-	"github.com/kyma-project/kyma/components/eventing-controller/pkg/deployment"
-	"github.com/kyma-project/kyma/components/eventing-controller/pkg/env"
 	"github.com/kyma-project/kyma/components/eventing-controller/utils"
+
+	"github.com/kyma-project/kyma/components/eventing-controller/pkg/object"
 )
-
-func TestApiRuleEqual(t *testing.T) {
-	svc := "svc"
-	port := uint32(9999)
-	host := "host"
-	isExternal := true
-	gateway := "foo.gateway"
-	labels := map[string]string{
-		"foo": "bar",
-	}
-	handler := &apigatewayv1beta1.Handler{
-		Name: "handler",
-	}
-	rule := apigatewayv1beta1.Rule{
-		Path: "path",
-		Methods: []string{
-			http.MethodPost,
-		},
-		AccessStrategies: []*apigatewayv1beta1.Authenticator{
-			{
-				Handler: handler,
-			},
-		},
-	}
-	apiRule := apigatewayv1beta1.APIRule{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "foo",
-			Namespace: "bar",
-			Labels:    labels,
-		},
-		Spec: apigatewayv1beta1.APIRuleSpec{
-			Service: &apigatewayv1beta1.Service{
-				Name:       &svc,
-				Port:       &port,
-				IsExternal: &isExternal,
-			},
-			Host:    &host,
-			Gateway: &gateway,
-			Rules:   []apigatewayv1beta1.Rule{rule},
-		},
-	}
-	testCases := map[string]struct {
-		prep   func() *apigatewayv1beta1.APIRule
-		expect bool
-	}{
-		"should be equal when svc, gateway, owner ref, rules are same": {
-			prep: func() *apigatewayv1beta1.APIRule {
-				apiRuleCopy := apiRule.DeepCopy()
-				return apiRuleCopy
-			},
-			expect: true,
-		},
-		"should be unequal when svc name is diff": {
-			prep: func() *apigatewayv1beta1.APIRule {
-				apiRuleCopy := apiRule.DeepCopy()
-				newSvcName := "new"
-				apiRuleCopy.Spec.Service.Name = &newSvcName
-				return apiRuleCopy
-			},
-			expect: false,
-		},
-		"should be unequal when svc port is diff": {
-			prep: func() *apigatewayv1beta1.APIRule {
-				apiRuleCopy := apiRule.DeepCopy()
-				newSvcPort := uint32(8080)
-				apiRuleCopy.Spec.Service.Port = &newSvcPort
-				return apiRuleCopy
-			},
-			expect: false,
-		},
-		"should be unequal when isExternal is diff": {
-			prep: func() *apigatewayv1beta1.APIRule {
-				apiRuleCopy := apiRule.DeepCopy()
-				newIsExternal := false
-				apiRuleCopy.Spec.Service.IsExternal = &newIsExternal
-				return apiRuleCopy
-			},
-			expect: false,
-		},
-		"should be unequal when gateway is diff": {
-			prep: func() *apigatewayv1beta1.APIRule {
-				apiRuleCopy := apiRule.DeepCopy()
-				newGateway := "new-gw"
-				apiRuleCopy.Spec.Gateway = &newGateway
-				return apiRuleCopy
-			},
-			expect: false,
-		},
-		"should be unequal when labels are diff": {
-			prep: func() *apigatewayv1beta1.APIRule {
-				apiRuleCopy := apiRule.DeepCopy()
-				newLabels := map[string]string{
-					"new-foo": "new-bar",
-				}
-				apiRuleCopy.Labels = newLabels
-				return apiRuleCopy
-			},
-			expect: false,
-		},
-		"should be unequal when path is diff": {
-			prep: func() *apigatewayv1beta1.APIRule {
-				apiRuleCopy := apiRule.DeepCopy()
-				newRule := rule.DeepCopy()
-				newRule.Path = "new-path"
-				apiRuleCopy.Spec.Rules = []apigatewayv1beta1.Rule{*newRule}
-				return apiRuleCopy
-			},
-			expect: false,
-		},
-		"should be unequal when methods are diff": {
-			prep: func() *apigatewayv1beta1.APIRule {
-				apiRuleCopy := apiRule.DeepCopy()
-				newRule := rule.DeepCopy()
-				newRule.Methods = []string{http.MethodOptions}
-				apiRuleCopy.Spec.Rules = []apigatewayv1beta1.Rule{*newRule}
-				return apiRuleCopy
-			},
-			expect: false,
-		},
-		"should be unequal when handlers are diff": {
-			prep: func() *apigatewayv1beta1.APIRule {
-				apiRuleCopy := apiRule.DeepCopy()
-				newRule := rule.DeepCopy()
-				newHandler := &apigatewayv1beta1.Handler{
-					Name: "foo",
-				}
-				newRule.AccessStrategies = []*apigatewayv1beta1.Authenticator{
-					{
-						Handler: newHandler,
-					},
-				}
-				apiRuleCopy.Spec.Rules = []apigatewayv1beta1.Rule{*newRule}
-				return apiRuleCopy
-			},
-			expect: false,
-		},
-		"should be unequal when OwnerReferences are diff": {
-			prep: func() *apigatewayv1beta1.APIRule {
-				apiRuleCopy := apiRule.DeepCopy()
-				newOwnerRef := metav1.OwnerReference{
-					APIVersion: "foo",
-					Kind:       "foo",
-					Name:       "foo",
-					UID:        "uid",
-				}
-				apiRuleCopy.OwnerReferences = []metav1.OwnerReference{
-					newOwnerRef,
-				}
-				return apiRuleCopy
-			},
-			expect: false,
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			testAPIRule := tc.prep()
-			if apiRuleEqual(&apiRule, testAPIRule) != tc.expect {
-				t.Errorf("expected output to be %t", tc.expect)
-			}
-		})
-	}
-}
-
-func TestEventingBackendEqual(t *testing.T) {
-	emptyBackend := eventingv1alpha1.EventingBackend{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "foo",
-			Namespace: "bar",
-		},
-		Spec: eventingv1alpha1.EventingBackendSpec{},
-	}
-
-	testCases := map[string]struct {
-		getBackend1    func() *eventingv1alpha1.EventingBackend
-		getBackend2    func() *eventingv1alpha1.EventingBackend
-		expectedResult bool
-	}{
-		"should be unequal if labels are different": {
-			getBackend1: func() *eventingv1alpha1.EventingBackend {
-				b := emptyBackend.DeepCopy()
-				b.Labels = map[string]string{"k1": "v1"}
-				return b
-			},
-			getBackend2: func() *eventingv1alpha1.EventingBackend {
-				return emptyBackend.DeepCopy()
-			},
-			expectedResult: false,
-		},
-		"should be equal if labels are the same": {
-			getBackend1: func() *eventingv1alpha1.EventingBackend {
-				b := emptyBackend.DeepCopy()
-				b.Labels = map[string]string{"k1": "v1"}
-				return b
-			},
-			getBackend2: func() *eventingv1alpha1.EventingBackend {
-				b := emptyBackend.DeepCopy()
-				b.Name = "bar"
-				b.Labels = map[string]string{"k1": "v1"}
-				return b
-			},
-			expectedResult: true,
-		},
-	}
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			if eventingBackendEqual(tc.getBackend1(), tc.getBackend2()) != tc.expectedResult {
-				t.Errorf("expected output to be %t", tc.expectedResult)
-			}
-		})
-	}
-}
 
 func TestEventingBackendStatusEqual(t *testing.T) {
 	testCases := []struct {
@@ -366,7 +153,7 @@ func TestEventingBackendStatusEqual(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
-			if IsBackendStatusEqual(tc.givenBackendStatus1, tc.givenBackendStatus2) != tc.wantResult {
+			if object.IsBackendStatusEqual(tc.givenBackendStatus1, tc.givenBackendStatus2) != tc.wantResult {
 				t.Errorf("expected output to be %t", tc.wantResult)
 			}
 		})
@@ -434,163 +221,8 @@ func Test_isSubscriptionStatusEqual(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
-			if gotEqualStatus := IsSubscriptionStatusEqual(tc.subscriptionStatus1, tc.subscriptionStatus2); tc.wantEqualStatus != gotEqualStatus {
+			if gotEqualStatus := object.IsSubscriptionStatusEqual(tc.subscriptionStatus1, tc.subscriptionStatus2); tc.wantEqualStatus != gotEqualStatus {
 				t.Errorf("The Subsciption Status are not equal, want: %v but got: %v", tc.wantEqualStatus, gotEqualStatus)
-			}
-		})
-	}
-}
-
-func TestPublisherProxyDeploymentEqual(t *testing.T) {
-	publisherCfg := env.PublisherConfig{
-		Image:          "publisher",
-		PortNum:        0,
-		MetricsPortNum: 0,
-		ServiceAccount: "publisher-sa",
-		Replicas:       1,
-		RequestsCPU:    "32m",
-		RequestsMemory: "64Mi",
-		LimitsCPU:      "64m",
-		LimitsMemory:   "128Mi",
-	}
-	natsConfig := env.NatsConfig{
-		EventTypePrefix: "prefix",
-		JSStreamName:    "kyma",
-	}
-	defaultNATSPublisher := deployment.NewNATSPublisherDeployment(natsConfig, publisherCfg)
-	defaultBEBPublisher := deployment.NewBEBPublisherDeployment(publisherCfg)
-
-	testCases := map[string]struct {
-		getPublisher1  func() *appsv1.Deployment
-		getPublisher2  func() *appsv1.Deployment
-		expectedResult bool
-	}{
-		"should be equal if same default NATS publisher": {
-			getPublisher1: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Name = "publisher1"
-				return p
-			},
-			getPublisher2: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Name = "publisher2"
-				return p
-			},
-			expectedResult: true,
-		},
-		"should be equal if same default BEB publisher": {
-			getPublisher1: func() *appsv1.Deployment {
-				p := defaultBEBPublisher.DeepCopy()
-				p.Name = "publisher1"
-				return p
-			},
-			getPublisher2: func() *appsv1.Deployment {
-				p := defaultBEBPublisher.DeepCopy()
-				p.Name = "publisher2"
-				return p
-			},
-			expectedResult: true,
-		},
-		"should be unequal if publisher types are different": {
-			getPublisher1: func() *appsv1.Deployment {
-				return defaultBEBPublisher.DeepCopy()
-			},
-			getPublisher2: func() *appsv1.Deployment {
-				return defaultNATSPublisher.DeepCopy()
-			},
-			expectedResult: false,
-		},
-		"should be unequal if publisher image changes": {
-			getPublisher1: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Template.Spec.Containers[0].Image = "new-publisher-img"
-				return p
-			},
-			getPublisher2: func() *appsv1.Deployment {
-				return defaultNATSPublisher.DeepCopy()
-			},
-			expectedResult: false,
-		},
-		"should be unequal if env var changes": {
-			getPublisher1: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Template.Spec.Containers[0].Env[0].Value = "new-value"
-				return p
-			},
-			getPublisher2: func() *appsv1.Deployment {
-				return defaultNATSPublisher.DeepCopy()
-			},
-			expectedResult: false,
-		},
-		"should be unequal if replicas changes": {
-			getPublisher1: func() *appsv1.Deployment {
-				replicas := int32(2)
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Replicas = &replicas
-				return p
-			},
-			getPublisher2: func() *appsv1.Deployment {
-				return defaultNATSPublisher.DeepCopy()
-			},
-			expectedResult: false,
-		},
-		"should be equal if spec annotations are nil and empty": {
-			getPublisher1: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Template.Annotations = nil
-				return p
-			},
-			getPublisher2: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Template.Annotations = map[string]string{}
-				return p
-			},
-			expectedResult: true,
-		},
-		"should be unequal if spec annotations changes": {
-			getPublisher1: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Template.Annotations = map[string]string{"key": "value1"}
-				return p
-			},
-			getPublisher2: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Template.Annotations = map[string]string{"key": "value2"}
-				return p
-			},
-			expectedResult: false,
-		},
-		"should be equal if spec Labels are nil and empty": {
-			getPublisher1: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Template.Labels = nil
-				return p
-			},
-			getPublisher2: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Template.Labels = map[string]string{}
-				return p
-			},
-			expectedResult: true,
-		},
-		"should be unequal if spec Labels changes": {
-			getPublisher1: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Template.Labels = map[string]string{"key": "value1"}
-				return p
-			},
-			getPublisher2: func() *appsv1.Deployment {
-				p := defaultNATSPublisher.DeepCopy()
-				p.Spec.Template.Labels = map[string]string{"key": "value2"}
-				return p
-			},
-			expectedResult: false,
-		},
-	}
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			if publisherProxyDeploymentEqual(tc.getPublisher1(), tc.getPublisher2()) != tc.expectedResult {
-				t.Errorf("expected output to be %t", tc.expectedResult)
 			}
 		})
 	}
